altered the startup order, refactored plugin initialisation to occur in bot rather than in the command dispatcher or handler
